+++ include/phidgets_spatial/spatial_ros_i.hpp
@@ -64,6 +64,8 @@ class SpatialRosI final : public rclcpp::Node
     void timerCallback();
     rclcpp::TimerBase::SharedPtr timer_;
     double publish_rate_;
+    std::string server_name_;
+    std::string server_ip_;
 
     rclcpp::Time ros_time_zero_;
     bool synchronize_timestamps_{true};
@@ -100,11 +102,19 @@ class SpatialRosI final : public rclcpp::Node
     double last_mag_y_{0.0};
     double last_mag_z_{0.0};
 
+    // Onboard orientation estimation results
+    double last_quat_w_;
+    double last_quat_x_;
+    double last_quat_y_;
+    double last_quat_z_;
+
     void publishLatest();
 
     void spatialDataCallback(const double acceleration[3],
                              const double angular_rate[3],
                              const double magnetic_field[3], double timestamp);
+    void spatialAlgorithmDataCallback(const double quaternion[4],
+                                      double timestamp);
     void attachCallback();
     void detachCallback();
 };